Compute for permission section
